teams: smoother members list (fixes #5626) (#5629)
