+++ app/src/main/java/org/ole/planet/myplanet/ui/team/teamMember/JoinedMemberFragment.kt
@@ -4,13 +4,30 @@ import android.content.res.Configuration
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import org.ole.planet.myplanet.base.BaseMemberFragment
+import org.ole.planet.myplanet.model.RealmMyTeam
 import org.ole.planet.myplanet.model.RealmMyTeam.Companion.getJoinedMember
 import org.ole.planet.myplanet.model.RealmNews
 import org.ole.planet.myplanet.model.RealmUserModel
 
 class JoinedMemberFragment : BaseMemberFragment() {
     override val list: List<RealmUserModel>
-        get() = getJoinedMember(teamId, mRealm)
+        get() {
+            val members = getJoinedMember(teamId, mRealm).toMutableList()
+            val leader = members.find { it.id == getTeamLeaderId() }
+            if (leader != null) {
+                members.remove(leader)
+                members.add(0, leader)
+            }
+            return members
+        }
+
+    private fun getTeamLeaderId(): String? {
+        val team = mRealm.where(RealmMyTeam::class.java)
+            .equalTo("teamId", teamId)
+            .equalTo("isLeader", true)
+            .findFirst()
+        return team?.userId
+    }
     override val adapter: RecyclerView.Adapter<*>
         get() = AdapterJoinedMember(requireActivity(), list.toMutableList(), mRealm, teamId)
